# backends/catalyst-interviewai/openapi-spec.yaml
swagger: '2.0'
info:
  title: InterviewAI Catalyst Backend API
  description: API for managing user-provided API keys, proxying AI requests, and RAG content processing for InterviewAI.
  version: '2.0.0'
# The host will be automatically generated when you deploy your API Gateway.
# Example: catalyst-gateway-xxxxxx-uc.a.run.app
# You can update this field later after deployment if needed.
host: your-gateway-id.uc.gateway.dev 
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json

# Disable all security - let Cloud Functions handle authentication
security: []
paths:
...
  /api/documents/parse-resume:
    options:
      summary: Handle CORS preflight requests for Parse Resume
      operationId: corsParseResume
      security: []
      x-google-backend:
        address: "%s" # Placeholder for ParseResume URL (9th)
        disable_auth: true
      responses:
        '200':
          description: CORS preflight response
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
    post:
      summary: Parses uploaded resume/document files
      description: Extracts text content from uploaded DOCX, PDF, or other document formats
      operationId: parseResume
      security: []
      x-google-backend:
        address: "%s" # Placeholder for ParseResume URL (10th)
        disable_auth: true
      responses:
        '200':
          description: Document parsed successfully
          schema:
            type: object
            properties:
              extractedText:
                type: string
              metadata:
                type: object
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  # ===== RAG ENDPOINTS =====
  /api/content/scrape:
    options:
      summary: Handle CORS preflight requests for Content Scraper
      operationId: corsContentScrape
      security: []
      x-google-backend:
        address: "%s" # Placeholder for ContentScraper URL (11th)
        disable_auth: true
      responses:
        '200':
          description: CORS preflight response
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
    post:
      summary: Scrapes content from YouTube videos or blog posts
      description: Extracts content, generates embeddings, and stores for RAG retrieval
      operationId: scrapeContent
      security: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - url
              - contentType
            properties:
              url:
                type: string
                description: URL to scrape (YouTube video or blog post)
              contentType:
                type: string
                enum: [youtube, blog]
                description: Type of content to scrape
              extractionOptions:
                type: object
                properties:
                  generateEmbeddings:
                    type: boolean
                    default: true
                  extractQuestions:
                    type: boolean
                    default: true
                  extractConcepts:
                    type: boolean
                    default: true
      x-google-backend:
        address: "%s" # Placeholder for ContentScraper URL (12th)
        disable_auth: true
      responses:
        '200':
          description: Content scraped successfully
          schema:
            $ref: '#/definitions/ScrapedContent'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Scraping failed
  /api/content/search:
    options:
      summary: Handle CORS preflight requests for Content Search
      operationId: corsContentSearch
      security: []
      x-google-backend:
        address: "%s" # Placeholder for ContentScraper URL (13th)
        disable_auth: true
      responses:
        '200':
          description: CORS preflight response
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
    get:
      summary: Searches scraped content by criteria
      description: Finds relevant content based on filters
      operationId: searchContent
      security: []
      parameters:
        - name: company
          in: query
          type: string
          description: Filter by target company
        - name: interviewType
          in: query
          type: string
          description: Filter by interview type
        - name: level
          in: query
          type: string
          description: Filter by target level
        - name: query
          in: query
          type: string
          description: Text search query
        - name: limit
          in: query
          type: integer
          default: 10
          description: Maximum number of results
      x-google-backend:
        address: "%s" # Placeholder for ContentScraper URL (14th)
        disable_auth: true
      responses:
        '200':
          description: Search results returned
          schema:
            type: object
            properties:
              results:
                type: array
                items:
                  $ref: '#/definitions/SearchResult'
              total:
                type: integer
        '401':
          description: Unauthorized
...
definitions:
...
  ScrapedContent:
    type: object
    properties:
      source:
        type: object
        properties:
          type:
            type: string
          url:
            type: string
          title:
            type: string
          author:
            type: string
          datePublished:
            type: string
      contentType:
        type: string
      interviewType:
        type: string
      targetLevel:
        type: string
      targetCompany:
        type: string
      content:
        type: object
        properties:
          questions:
            type: array
            items:
              type: object
          concepts:
            type: array
            items:
              type: object
          tips:
            type: array
            items:
              type: object
          fullTranscript:
            type: string
      qualityScore:
        type: number
      embeddings:
        type: object
  SearchResult:
    type: object
    properties:
      id:
        type: string
      content:
        type: string
      source:
        type: string
      title:
        type: string
      score:
        type: number
      contentType:
        type: string
      metadata:
        type: object
...